backend/requirements.txt
fastapi
uvicorn[standard]
python-pptx
anthropic
sqlalchemy
python-multipart
python-dotenv
pydantic
